modify(order): 주문 상품 아이디 목록 조회 proto 수정 (데이터 타입)

+++ order/src/main/java/com/order/applicatoin/port/in/command/FindOrderProductIdsCommand.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record FindOrderProductIdsCommand(
     Long customerId,
-    long limit
+    int limit
 ) {
 
 }
